# functional courtyards calculation