--- a/momepy/__init__.py
+++ b/momepy/__init__.py
@@ -11,6 +11,7 @@ from .functional._dimension import *
 from .functional._distribution import *
 from .functional._diversity import *
 from .functional._elements import *
+from .functional._intensity import *
 from .functional._shape import *
 from .graph import *
 from .intensity import *