Refactoring courtyards calculation to fit with the functional API and to speed it up.