@u3ks 2024-04-25
Timings:

new time: 11.6 s ± 191 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
old time: 43.8 s ± 635 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)